% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_package.R
\docType{package}
\name{addin_package}
\alias{addin_package}
\alias{addin_package-package}
\title{RStudio Addins}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{findAndReplaceAddin}()} and \code{\link{insertInAddin}()}
for two examples of R functions that are exported.
}

