% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCommas.R, R/insertAddressAddin.R
\name{insertAddressAddin}
\alias{insertAddressAddin}
\title{Insert \%>\%.}
\usage{
insertAddressAddin()

insertAddressAddin()
}
\description{
Call this function as an addin to insert \code{ \%>\% } at the cursor position.
helpful for using dplyr

Call this function as an addin to insert \code{ \%>\% } at the cursor position.
helpful for using dplyr
}

