% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertArrowAddin.R
\name{insertArrowAddin}
\alias{insertArrowAddin}
\title{insertArrowAddin}
\usage{
insertArrowAddin()
}
\value{
Inserts an <- into the code.
}
\description{
addins: insertArrowAddin
Call this function as an addin to insert <- at the cursor position.
}
\examples{

insertArrowAddin()


}
\keyword{insertArrowAddin}
