% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertComma.R
\name{insertComma}
\alias{insertComma}
\title{Insert comma}
\usage{
insertComma()
}
\description{
Moves out of auto-completed quotation marks or parenthesis and
inserts a comma and space at the new cursor position
}
