% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDoubleAmpersand.R
\name{insertDoubleAmpersand}
\alias{insertDoubleAmpersand}
\title{insertDoubleAmpersand}
\usage{
insertDoubleAmpersand()
}
\value{
Inserts an && into the code.
}
\description{
addins: insertDoubleAmpersand
}
\details{
Call this function as an addin to insert && at the cursor position.
}
\examples{

insertDoubleAmpersand()

}
\keyword{insertDoubleAmpersand}
