% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDoublEqual.R
\name{insertDoubleEqual}
\alias{insertDoubleEqual}
\title{Insert ==}
\usage{
insertDoubleEqual()
}
\description{
Call this function as an addin to insert \code{ == } at the cursor position.
}
\examples{

==

}
