% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDoubleOr.R
\name{insertDoubleOr}
\alias{insertDoubleOr}
\title{insertDoubleOr}
\usage{
insertDoubleOr()
}
\value{
Inserts an || into the code.
}
\description{
addins: insertDoubleOr
}
\details{
Call this function as an addin to insert || at the cursor position.
}
\examples{

insertDoubleOr()

}
\keyword{insertDoubleOr}
