% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGreaterEqual.R
\name{insertGreaterEqual}
\alias{insertGreaterEqual}
\title{insertGreaterEqual}
\usage{
insertGreaterEqual()
}
\value{
Inserts an => into the code.
}
\description{
addins: insertGreaterEqual()
}
\details{
Call this function as an addin to insert => at the cursor position.
}
\examples{

insertGreaterEqual()

}
\keyword{insertGreaterEqual}
