% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertInAddin.R
\name{insertInAddin}
\alias{insertInAddin}
\title{insertInAddin}
\usage{
insertInAddin()
}
\value{
Inserts an %in%  into the code.
}
\description{
addins: insertInAddin
}
\details{
Call this function as an addin to insert %in% at the cursor position.
}
\examples{

insertInAddin()

}
\keyword{insertInAddin}
