% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLessEqual.R
\name{insertLessEqual}
\alias{insertLessEqual}
\title{insertLessEqual}
\usage{
insertLessEqual()
}
\value{
Inserts an <= into the code.
}
\description{
addins: insertLessEqual
}
\details{
Call this function as an addin to insert <= at the cursor position.
}
\examples{
insertLessEqual()

}
\keyword{insertLessEqual}
