% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertNotequal.R
\name{insertNotEqual}
\alias{insertNotEqual}
\title{insertNotEqual}
\usage{
insertNotEqual()
}
\value{
Inserts an != into the code.
}
\description{
addins: insertNotEqual
}
\details{
Call this function as an addin to insert != at the cursor position.
}
\examples{
 

insertNotEqual()

}
\keyword{insertNotEqual}
