% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPipe.R
\name{insertPipe}
\alias{insertPipe}
\title{insertPipe}
\usage{
insertPipe()
}
\value{
Inserts an %>% into the code.
}
\description{
addins: insertPipe
}
\details{
Call this function as an addin to insert %>% at the cursor position.
}
\examples{
library(rstudioapi)
library(addins)

insertPipe()

}
\keyword{insertPipe}
