% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPipeAddin.R
\name{insertPipeAddin}
\alias{insertPipeAddin}
\title{Insert \%>\%.}
\usage{
insertPipeAddin()
}
\description{
Call this function as an addin to insert \code{ \%>\% } at the cursor position.
helpful for using dplyr
}
