% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertArrowin.R
\name{insertArrowAddin}
\alias{insertArrowAddin}
\title{Insert <-}
\usage{
insertArrowAddin()
}
\description{
Call this function as an addin to insert \code{ <- } at the cursor position.
}
\examples{

  <-

}
