% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDoubleAmpersand.R
\name{insertDoubleAmpersand}
\alias{insertDoubleAmpersand}
\title{Insert &&}
\usage{
insertDoubleAmpersand()
}
\description{
Call this function as an addin to insert \code{ && } at the cursor position.
}
\examples{

&&

}
