% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDoublEqual.R
\name{insertDoubleEqual}
\alias{insertDoubleEqual}
\title{insertDoubleEqual}
\usage{
insertDoubleEqual()
}
\value{
Inserts an == into the code.
}
\description{
addins: insertDoubleEqual
}
\details{
Call this function as an addin to insert == at the cursor position.
}
\examples{

insertDoubleEqual()

}
\keyword{insertDoubleEqual}
