% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGreaterEqual.R
\name{insertGreaterEqual}
\alias{insertGreaterEqual}
\title{Insert =>}
\usage{
insertGreaterEqual()
}
\description{
Call this function as an addin to insert \code{ <= } at the cursor position.
}
\examples{

=>

}
